+++ src/main/java/nextstep/subway/section/domain/SectionRepository.java
@@ -1,4 +1,4 @@
-package nextstep.subway.line.domain;
+package nextstep.subway.section.domain;
 
 import org.springframework.data.jpa.repository.JpaRepository;
