+++ src/main/java/nextstep/subway/section/dto/SectionResponse.java
@@ -1,0 +1,4 @@
+package nextstep.subway.section.dto;
+
+public class SectionResponse {
+}
