+++ src/main/java/nextstep/subway/section/ui/SectionController.java
@@ -1,0 +1,8 @@
+package nextstep.subway.section.ui;
+
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+public class SectionController {
+    
+}
