+++ src/main/java/nextstep/subway/station/dto/StationRequest.java
@@ -11,10 +11,6 @@ public class StationRequest {
         this.name = name;
     }
 
-    public String getName() {
-        return name;
-    }
-
     public Station toStation() {
         return new Station(name);
     }
